Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/workers/nested_worker_terminate_from_document-expected.txt


FAIL Test terminating a chain of nested workers by calling terminate() from the owning document assert_equals: expected "Pass" but got "Fail: ReferenceError: Can't find variable: Worker"
